Introduced ip replacement mappings for behind NAT depots
ibp_server sends bind ip in the CAPs to the client. Many clients are
programmed to use that ip for successive communication. i.e. ibp_alloc
and ibp_store.

But these behind NAT ips may not be rechable to client. So this mapping
option, allows us to specify public ips for default interfaces. If set,
ibp_server will send that public ip in CAP only when we client contacts on
default interfaces.